+++ betty/jinja_tests/image_media_type_supported.py
