+++ betty/tests/jinja_tests/test_image_media_type_supported.py
