+++ pyproject.toml
@@ -248,12 +248,12 @@ betty = 'betty.console:main_standalone'
 'webpack-entry-point-js' = 'betty.plugins.jinja_filter.webpack_entry_point_js:WebpackEntryPointJs'
 
 [project.entry-points.'betty.jinja_test']
-'has-file-references' = 'betty.plugins.jinja_test.has_file_references:HasFileReferences'
-'image-media-type-supported' = 'betty.plugins.jinja_test.image_media_type_supported:ImageMediaTypeSupported'
-'linked-data-dumpable' = 'betty.plugins.jinja_test.linked_data_dumpable:LinkedDataDumpable'
-'persistent-entity-id' = 'betty.plugins.jinja_test.persistent_entity_id:PersistentEntityId'
-'public' = 'betty.plugins.jinja_test.public:Public'
-'plugin' = 'betty.plugins.jinja_test.plugin:Plugin.discover'
+'has-file-references' = 'betty.jinja_tests.has_file_references:HasFileReferences'
+'image-media-type-supported' = 'betty.jinja_tests.image_media_type_supported:ImageMediaTypeSupported'
+'linked-data-dumpable' = 'betty.jinja_tests.linked_data_dumpable:LinkedDataDumpable'
+'persistent-entity-id' = 'betty.jinja_tests.persistent_entity_id:PersistentEntityId'
+'public' = 'betty.jinja_tests.public:Public'
+'plugin' = 'betty.jinja_tests.plugin:Plugin.discover'
 
 
 [project.entry-points.'betty.js_resource']
